requirements.txt
flask==2.3.3
flask
psycopg2-binary
